Добавил UiState и ViewModel для экрана ProductCardScreen
